refactor(gateway): split reload handlers
